[Chore] #129 - 폴더 구조 변경
