+++ Smeem-iOS/Smeem-iOS/Global/Extensions/UIViewController+.swift
@@ -128,7 +128,7 @@ extension UIViewController {
     }
     
     func presentOnboardingAcceptVC() {
-        let bottomSheetVC = LoginBottomSheetViewController(loginManager: LoginManagerImpl(loginService: LoginServiceImpl(requestable: APIService())))
+        let bottomSheetVC = LoginBottomSheetViewController(loginManager: LoginManagerImpl(loginService: LoginServiceImpl(requestable: RequestImpl())))
         let onboardingAcceptVC = UserNicknameViewController()
         onboardingAcceptVC.userPlanRequest = bottomSheetVC.userPlanRequest
         let _ = UINavigationController(rootViewController: onboardingAcceptVC)
