+++ Smeem-iOS/Smeem-iOS/Network/DataModel/BadgeModel.swift
@@ -1,14 +1,13 @@
 //
-//  BadgeListModel.swift
+//  BadgeModel.swift
 //  Smeem-iOS
 //
 //  Created by 황찬미 on 2023/06/28.
 //
 
 import Foundation
-import UIKit
 
-// MARK: - Dummy
+// MARK: - Badge Dummy
 
 struct DummyModel {
     func dummyBadgeData() -> [[(String, String)]] {
@@ -21,7 +20,7 @@ struct DummyModel {
     }
 }
 
-// MARK: BadgeListResponse
+// MARK: BadgeListResponse (myPage)
 
 struct BadgeListResponse: Codable {
     let badges: [BadgesListArray]
@@ -40,4 +39,26 @@ struct BadgesListArray: Codable {
 
 extension BadgesListArray {
     static let empty = [BadgesListArray(id: 0, name: "", type: "", imageURL: "")]
+}
+
+// MARK: Badges (onboarding, diary -> home)
+
+struct Badges: Codable {
+    let id: Int
+    let name: String
+    let type: String
+    let imageUrl: String
+}
+
+// MARK: Badge (myPage)
+
+struct Badge: Codable {
+    let id: Int
+    let name, type: String
+    let imageURL: String
+
+    enum CodingKeys: String, CodingKey {
+        case id, name, type
+        case imageURL = "imageUrl"
+    }
 }
