+++ Smeem-iOS/Smeem-iOS/Network/DataModel/Home/HomeDiaryResponse.swift
