+++ Smeem-iOS/Smeem-iOS/Network/DataModel/NicknameModel.swift
@@ -1,0 +1,12 @@
+//
+//  NicknameModel.swift
+//  Smeem-iOS
+//
+//  Created by 황찬미 on 2023/09/27.
+//
+
+import Foundation
+
+struct NicknameValidResponse: Codable {
+    let isExist: Bool
+}
