+++ Smeem-iOS/Smeem-iOS/Network/EndPoint/LoginEndPoint.swift
@@ -61,7 +61,6 @@ extension LoginEndPoint: BaseEndPoint {
     func makeUrlRequest() -> NetworkRequest {
         return NetworkRequest(path: path,
                               httpMethod: httpMethod,
-                              query: query,
                               body: requestBody,
                               headers: header)
     }
