+++ Smeem-iOS/Smeem-iOS/Network/Manager/SplashManagerImpl.swift
