+++ Smeem-iOS/Smeem-iOS/Presentation/Home/HomeViewController.swift
@@ -221,7 +221,7 @@ final class HomeViewController: UIViewController {
     }
     
     @objc func myPageButtonDidTap(_ sender: UIButton) {
-        let myPageVC = MyPageViewController(myPageManager: MyPageManagerImpl(myPageService: MyPageServiceImpl(requestable: APIService())), editPushManager: MyPageEditManagerImpl(myPageEditService: MyPageEditServiceImpl(requestable: APIService())))
+        let myPageVC = MyPageViewController(myPageManager: MyPageManagerImpl(myPageService: MyPageServiceImpl(requestable: RequestImpl())), editPushManager: MyPageEditManagerImpl(myPageEditService: MyPageEditServiceImpl(requestable: RequestImpl())))
         self.navigationController?.pushViewController(myPageVC, animated: true)
     }
