--- Smeem-iOS/Smeem-iOS/Presentation/Onboarding/Goal/Model/PlanListResponse.swift
@@ -1,16 +1,0 @@
-//
-//  PlanListResponse.swift
-//  Smeem-iOS
-//
-//  Created by 황찬미 on 2023/07/01.
-//
-
-import Foundation
-
-struct PlanListResponse: Codable {
-    let goals: [Plan]
-}
-
-struct Plan: Codable {
-    let goalType, name: String
-}
