+++ Smeem-iOS/Smeem-iOS/Presentation/Onboarding/Goal/VC/GoalViewController.swift
@@ -85,7 +85,7 @@ extension GoalViewController: NextButtonDelegate {
             howOnboardingVC.tempTarget = selectedGoalLabel
             self.navigationController?.pushViewController(howOnboardingVC, animated: true)
         case .myPage:
-            let howOnboardingVC = EditGoalViewController(editGoalManager: MyPageEditManagerImpl(myPageEditService: MyPageEditServiceImpl(requestable: APIService())))
+            let howOnboardingVC = EditGoalViewController(editGoalManager: MyPageEditManagerImpl(myPageEditService: MyPageEditServiceImpl(requestable: RequestImpl())))
             howOnboardingVC.tempTarget = selectedGoalLabel
             self.navigationController?.pushViewController(howOnboardingVC, animated: true)
         }
