--- Smeem-iOS/Smeem-iOS/Presentation/Onboarding/How/Model/DetailPlanListResponse.swift
@@ -1,14 +1,0 @@
-//
-//  DetailPlanListResponse.swift
-//  Smeem-iOS
-//
-//  Created by 황찬미 on 2023/07/01.
-//
-
-import Foundation
-
-struct DetailPlanListResponse:Codable {
-    let name: String
-    let way: String
-    let detail: String
-}
